requirements.txt
asgiref==3.8.1
asttokens==3.0.0
async-timeout==5.0.1
beautifulsoup4==4.13.3
black==25.1.0
celery==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
decorator==5.2.1
Django==4.2.20
django-cors-headers==4.7.0
django-extensions==3.2.3
django-filter==24.3
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-gis==1.1
environs==14.1.1
exceptiongroup==1.2.2
executing==2.2.0
fake-useragent==2.1.0
geopandas==1.0.1
gunicorn==23.0.0
idna==3.10
ipython==8.18.1
jedi==0.19.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
mypy-extensions==1.0.0
numpy==2.0.2
packaging==24.2
pandas==2.2.3
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
platformdirs==4.3.7
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.1
pyogrio==0.10.0
pyproj==3.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
redis==5.2.1
requests==2.32.3
shapely==2.0.7
six==1.17.0
soupsieve==2.6
sqlparse==0.5.3
stack-data==0.6.3
tomli==2.2.1
traitlets==5.14.3
tzdata==2025.2
urllib3==2.3.0
uuid==1.30
wcwidth==0.2.13
